requirements.txt
fastapi
jinja2
psycopg2-binary
python-dotenv
loguru